+++ tests/projects/c++/modules/stl_headerunit/xmake.lua
@@ -1,7 +1,11 @@
 add_rules("mode.release", "mode.debug")
 set_languages("c++20")
 
-target("stl_headerunit")
+-- clang 编译 module 报错，暂时禁用
+set_config("cc", "gcc")
+set_config("cxx", "g++")
+
+target("stl_headerunit", function()
     set_kind("binary")
     add_files("src/*.cpp", "src/*.mpp")
-
+end)